Fixes @Inventitech's comments
Refs #224

+++ WatchDogCore/WatchDog/src/nl/tudelft/watchdog/core/logic/ui/events/EditorEvent.java
@@ -7,21 +7,21 @@ public class EditorEvent extends WatchDogEvent {
 	/** Serial ID. */
 	private static final long serialVersionUID = 1L;
 	
-	private long modifiedChars;
+	private int modifiedChars;
 
 	/** Constructor. */
 	public EditorEvent(Object editor, EventType type) {
 		super(editor, type);
-		modifiedChars = 0L;
+		modifiedChars = 0;
 	}
 	
 	/** Sets the number of characters that have been modified by this event. */
-	public void setModCount(long nrOfModifiedChars) {
+	public void setModCount(int nrOfModifiedChars) {
 		this.modifiedChars = nrOfModifiedChars;
 	}
 	
 	/** Gets the number of characters that have been modified by this event. */
-	public long getModCount() {
+	public int getModCount() {
 		return this.modifiedChars;
 	}
